+++ packages/snap-controller/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.30.2](https://github.com/searchspring/snap/compare/v0.30.1...v0.30.2) (2022-07-06)
+
+
+### Bug Fixes
+
+* fix lastStringyParams logic ([9cab5cb](https://github.com/searchspring/snap/commit/9cab5cbbc16529928ae6f55e80a315ad82bbe04d))
+* **searchcontroller:** remove search.redirectResponse and personalization when comparing params ([93b40a2](https://github.com/searchspring/snap/commit/93b40a2f9b0c43fe4aefb6ccb5a9cb9c74b76d9d))
+
+
+
+
+
 ## [0.30.1](https://github.com/searchspring/snap/compare/v0.30.0...v0.30.1) (2022-07-02)
 
 **Note:** Version bump only for package @searchspring/snap-controller