+++ packages/snap-store-mobx/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.30.2](https://github.com/searchspring/snap/compare/v0.30.1...v0.30.2) (2022-07-06)
+
+
+### Bug Fixes
+
+* **mobx-store-finder:** removing the `data` property to prevent future access problems ([b523369](https://github.com/searchspring/snap/commit/b52336973614c07a28a93b1fcf5ce465689c0cb2))
+
+
+
+
+
 ## [0.30.1](https://github.com/searchspring/snap/compare/v0.30.0...v0.30.1) (2022-07-02)
 
 **Note:** Version bump only for package @searchspring/snap-store-mobx